fix: keep pure-function spec imports on leaf modules (#1564)
The upload-card color spec imported IngestionResult through the
backoffice store, whose chain reaches src/i18n's Vite-only
import.meta.glob and crashes the Playwright node runner in CI.
IngestionState/IngestionResult move to src/constant/ingestion (leaf,
re-exported from the store so existing importers keep one path) and
resolveDataButtonColor moves to src/composables/uploadCardColor —
mirroring the pipelinePhaseLabel leaf-module pattern.
